Promote `trace` module to public API.
